redux + navigation intial config
