+++ assets/index.ts
@@ -9,6 +9,9 @@ export type AssetImages = {
     이안정: ImageSourcePropType;
     카와카츠: ImageSourcePropType;
     야키토리나루토: ImageSourcePropType;
+    placeholder: ImageSourcePropType;
+    default_profile: ImageSourcePropType;
+    kakao_login_medium_narrow: ImageSourcePropType;
   };
 };
 
@@ -20,6 +23,9 @@ const assets: AssetImages = {
     이안정: require('./image/4.png'),
     카와카츠: require('./image/5.png'),
     야키토리나루토: require('./image/6.png'),
+    placeholder: require('./image/placeholder.png'),
+    default_profile: require('./image/default_profile.png'),
+    kakao_login_medium_narrow: require('./image/kakao_login_medium_narrow.png'),
   },
 };
