+++ package.json
@@ -11,8 +11,10 @@
   },
   "dependencies": {
     "@gorhom/bottom-sheet": "^4.4.7",
+    "@react-native-async-storage/async-storage": "^1.19.0",
     "@react-navigation/native": "^6.1.6",
     "@react-navigation/stack": "^6.3.16",
+    "@reduxjs/toolkit": "^1.9.5",
     "react": "18.2.0",
     "react-native": "0.71.8",
     "react-native-elements": "^3.4.3",
@@ -25,8 +27,11 @@
     "react-native-reanimated": "^2.17.0",
     "react-native-safe-area-context": "^4.5.3",
     "react-native-safe-area-view": "^1.1.1",
+    "react-native-screens": "^3.22.1",
     "react-native-status-bar-height": "^2.6.0",
-    "react-native-vector-icons": "^9.2.0"
+    "react-native-vector-icons": "^9.2.0",
+    "react-redux": "^8.1.1",
+    "redux-persist": "^6.0.0"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0",
