+++ src/screens/ProfileMain.tsx
@@ -11,9 +11,7 @@ import {
   TextInput,
 } from 'react-native';
 import Ionicons from 'react-native-vector-icons/Ionicons';
-
-// implement profile picture later
-const PROFILE_PICTURE = '../assets/image/default_profile.png';
+import assets from '../../assets';
 
 export default function Settings() {
   const [isEdit, setIsEdit] = React.useState(false);
@@ -35,7 +33,7 @@ export default function Settings() {
               <View>
                 <Image
                   alt="Profile picture"
-                  source={require(PROFILE_PICTURE)}
+                  source={assets.images.default_profile}
                   style={styles.profileImage}
                 />
               </View>
