+++ src/types/navigation.ts
@@ -1,0 +1,15 @@
+export type ScreenParamList = RootStackParamList;
+
+export type RootStackParamList = {
+  LoginMain: undefined;
+  SettingsMain: undefined;
+
+  //BottomTab
+  MapMain: undefined;
+  ProfileMain: undefined;
+};
+
+export type BottomTabParamList = {
+  MapMain: undefined;
+  ProfileMain: undefined;
+};
